Revert "Clean and fixes to smart contract"
