--- .github/workflows/docker-test.yml
@@ -1,20 +1,0 @@
-name: Smart Contracts
-
-on:
-  push:
-    branches: [ develop ]
-  pull_request:
-    branches: [ develop ]
-
-jobs:
-
-  build:
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v2
-    - name: Build the Docker image
-      run: docker build . --file Dockerfile.test --tag dockertest
-    - name: Run test on docker
-      uses: ./ # Uses an action in the root directory
-      id: test
-
