--- action.yml
@@ -1,9 +1,0 @@
-# action.yml
-name: 'Docker test'
-description: 'test with docker'
-outputs:
-  time: # id of output
-    description: 'The time we greeted you'
-runs:
-  using: 'docker'
-  image: 'dockertest'
