+++ contracts/Migrations.sol
@@ -1,7 +1,4 @@
-// contracts/GLDToken.sol
-// SPDX-License-Identifier: MIT
-
-pragma solidity >=0.7.6;
+pragma solidity ^0.7.6;
 
 contract Migrations {
   address public owner = msg.sender;
