+++ contracts/_deprecated/forwarder.sol
@@ -1,5 +1,4 @@
-// SPDX-License-Identifier: GPL-2.0-or-later
-pragma solidity >=0.7.0;
+pragma solidity ^0.7.0;
 
 // import "../IERC20.sol";
 import "./relayer.sol";
@@ -15,7 +14,7 @@ contract Forwarder {
     /// @dev Fallback function allows to deposit ether.
     receive() external payable {
         //wallet.transfer(msg.value);
-        Relayer(wallet).processDeposit{value:msg.value}();
+        Relayer(wallet).processDeposit{value:msg.value}(msg.sender);
 
     }
 }
