+++ contracts/core/wallet_data.sol
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
-pragma solidity >=0.7.6;
+pragma solidity ^0.7.6;
 pragma experimental ABIEncoderV2;
 
 library Core {
