+++ contracts/ens/ens_resolver.sol
@@ -1,5 +1,4 @@
-// SPDX-License-Identifier: MIT
-pragma solidity >=0.7.6;
+pragma solidity ^0.7.6;
 
 /**
  * @title EnsResolver
@@ -8,5 +7,5 @@ pragma solidity >=0.7.6;
 abstract contract Resolver {
     function supportsInterface(bytes4 interfaceID) virtual public pure returns (bool);
     function addr(bytes32 node) virtual public view returns (address);
-    function setAddr(bytes32 node, address addr_) virtual public;
+    function setAddr(bytes32 node, address addr) virtual public;
 }
