+++ contracts/external/create2.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: UNLICENSED
 // Taken from: https://github.com/OpenZeppelin/openzeppelin-contracts/blob/970f687f04d20e01138a3e8ccf9278b1d4b3997b/contracts/utils/Create2.sol
 
-pragma solidity >=0.7.0;
+pragma solidity ^0.7.0;
 
 /**
  * @dev Helper to make usage of the `CREATE2` EVM opcode easier and safer.
