+++ contracts/features/daily_limit.sol
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
-pragma solidity >=0.7.6;
+pragma solidity ^0.7.6;
 import "../core/wallet_data.sol";
 
 library DailyLimit {
@@ -29,7 +29,7 @@ library DailyLimit {
     /// @dev Returns maximum withdraw amount.
     /// @return Returns amount.
     function calcMaxWithdraw(Core.Wallet storage _wallet)
-        public view
+        public
         returns (uint)
     {
         if (block.timestamp > _wallet.lastDay + 24 hours)
