+++ contracts/features/guardians.sol
@@ -1,7 +1,7 @@
 
 // SPDX-License-Identifier: GPL-3.0-only
 
-pragma solidity >=0.7.6;
+pragma solidity ^0.7.6;
 
 import "../core/wallet_data.sol";
