--- contracts/test/erc20_sample.sol
@@ -1,11 +1,0 @@
-// contracts/GLDToken.sol
-// SPDX-License-Identifier: MIT
-pragma solidity >=0.7.6;
-
-import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
-
-contract ERC20SampleToken is ERC20 {
-    constructor(uint256 initialSupply) ERC20("Gold", "GLD") {
-        _mint(msg.sender, initialSupply);
-    }
-}
