+++ package.json
@@ -5,20 +5,15 @@
   },
   "name": "@hashmesan/smartvault",
   "homepage": "https://github.com/hashmesan/harmony-totp",
-  "version": "0.1.4",
+  "version": "0.1.2",
   "license": "GPL-3.0",
-  "bin": {
-    "smartvault": "cli/smartvault.js"
-  },
+  "bin": { "smartvault": "cli/smartvault.js"},
   "dependencies": {
     "@harmony-js/core": "^0.1.57",
     "@harmony-js/crypto": "^0.1.56",
-    "@openzeppelin/contracts": "^4.2.0",
     "@truffle/contract": "^4.3.21",
     "@truffle/hdwallet-provider": "^1.4.1",
     "axios": "^0.21.1",
-    "bn-chai": "^1.0.1",
-    "bn.js": "^5.2.0",
     "chai": "^4.3.4",
     "chai-as-promised": "^7.1.1",
     "commander": "8",
