+++ relayer/test/test_ipfs.js
@@ -6,7 +6,7 @@ var wallet = require("../../lib/wallet.js");
 
 const env = "development";
 
-describe.skip('test IPFS()', function () {
+describe('test IPFS()', function () {
 //   it('test add', async function () {
 //       const text = "abcdef";
 //       const res = await ipfs.add(text);
