--- secrets.js.template
@@ -1,7 +1,0 @@
-
-module.exports = { 
-    // example keys from ganache-cli
-    rinkeby: "a1e50b72f59f06db18b511851836df0f2c646e163e74b50cde8b31b90c1c11dc",
-    harmonytestnet: "a1e50b72f59f06db18b511851836df0f2c646e163e74b50cde8b31b90c1c11dc",
-    harmonymainnet: "a1e50b72f59f06db18b511851836df0f2c646e163e74b50cde8b31b90c1c11dc"
-};
