+++ test/guardians_test.js
@@ -13,39 +13,23 @@ contract("Guardians", accounts => {
 
     it("should add guardian", async () => {
         var tmpWallet = web3.eth.accounts.create();
-        var { root, leaves, wallet } = await commons.createWallet(
-            ethers.constants.AddressZero,
-            ["", ""],
-            accounts[0],
-            8,
-            web3.utils.toWei("0.1", "ether"),
-            tmpWallet.address,
-            tmpWallet.address,
-            0);
-        
-        await wallet.addGuardian(tmpWallet.address);
+        var {startCounter, root, leaves, wallet} = await commons.createWallet(timeOffset, DURATION, 16, tmpWallet.address);
+        var proof = await commons.getTOTPAndProof(leaves, timeOffset, DURATION);
+        await wallet.addGuardian(tmpWallet.address, proof[0], proof[1]);
         var guardians = await wallet.getGuardians();
-        //console.log(guardians);
+        console.log(guardians);
         assert.equal(guardians.length, 1);
     })
 
     it("should remove guardian", async () => {
         var tmpWallet = web3.eth.accounts.create();
-        var { root, leaves, wallet } = await commons.createWallet(
-            ethers.constants.AddressZero,
-            ["", ""],
-            accounts[0],
-            8,
-            web3.utils.toWei("0.1", "ether"),
-            tmpWallet.address,
-            tmpWallet.address,
-            0);
-
-        await wallet.addGuardian(tmpWallet.address);
-        await wallet.revokeGuardian(tmpWallet.address);
+        var {startCounter, root, leaves, wallet} = await commons.createWallet(timeOffset, DURATION, 16, tmpWallet.address);
+        var proof = await commons.getTOTPAndProof(leaves, timeOffset, DURATION);
+        await wallet.addGuardian(tmpWallet.address, proof[0], proof[1]);
+        await wallet.revokeGuardian(tmpWallet.address, proof[0], proof[1]);
 
         var guardians = await wallet.getGuardians();
-        //console.log(guardians);
+        console.log(guardians);
         assert.equal(guardians[0], "0x0000000000000000000000000000000000000000");
     })    
 })
