+++ test/otp_wallet_test.js
@@ -87,7 +87,7 @@ contract("OTPWallet", accounts => {
         var {root, leaves, wallet} = await commons.createWallet(
             ethers.constants.AddressZero,
             ["",""],
-            tmpWallet.address,
+            accounts[0] ,
             8, 
             web3.utils.toWei("100", "ether"),
             tmpWallet.address, 
@@ -97,7 +97,7 @@ contract("OTPWallet", accounts => {
         var wallet2 = await commons.createWallet(
             ethers.constants.AddressZero,
             ["",""],
-            tmpWallet.address,
+            accounts[0] ,
             8, 
             web3.utils.toWei("100", "ether"),
             tmpWallet.address, 
@@ -106,7 +106,13 @@ contract("OTPWallet", accounts => {
 
         await web3.eth.sendTransaction({from: accounts[0], to: wallet.address, value: web3.utils.toWei("2", "ether")});
 
-        // send to the other smart wallet        
+        // send to the other smart wallet
+        await wallet.makeTransfer(wallet2.wallet.address, web3.utils.toWei("0.5", "ether"));
+        var newBalance = await web3.eth.getBalance(wallet2.wallet.address);
+        console.log("Balance 1=", newBalance);
+        assert.equal(newBalance, web3.utils.toWei(".5", "ether"), "transfer is correct");        
+
+
         const methodData = wallet.contract.methods.makeTransfer(wallet2.wallet.address, web3.utils.toWei("0.1234", "ether")).encodeABI();
         const nonce = await commons.getNonceForRelay();
         const gasLimit = 100000;
@@ -124,27 +130,8 @@ contract("OTPWallet", accounts => {
         );
 
         await wallet.executeMetaTx(methodData, sigs, nonce, 0, gasLimit, ethers.constants.AddressZero, feeWallet.address);       
-        var newBalance = await web3.eth.getBalance(wallet2.wallet.address);
-        assert.strictEqual(newBalance, web3.utils.toWei("0.1234", "ether"))                  
+        newBalance = await web3.eth.getBalance(wallet2.wallet.address);
         console.log("Balance 2 =", newBalance);
-
-        // try with wrong key
-        var sigs = await commons.signOffchain2(
-            [feeWallet],
-            wallet.address,
-            0,
-            methodData,
-            0,
-            nonce,
-            0,
-            gasLimit,
-            ethers.constants.AddressZero,
-            feeWallet.address
-        );
-
-        var tx = wallet.executeMetaTx(methodData, sigs, nonce, 0, gasLimit, ethers.constants.AddressZero, feeWallet.address);
-        await truffleAssert.reverts(tx, "RM: Invalid signatures -- Reason given: RM: Invalid signatures");       
-
     });
 
     async function createFactoryWallet(owner, salt) {
@@ -181,7 +168,7 @@ contract("OTPWallet", accounts => {
 
         return walletAddrComputed;        
     }
-    it("should transfer between smartvaults via factory", async()=>{
+    it.only("should transfer between smartvaults via factory", async()=>{
         var tmpWallet = web3.eth.accounts.create();
         var tmpWallet2 = web3.eth.accounts.create();
 
@@ -210,6 +197,8 @@ contract("OTPWallet", accounts => {
         );
 
         var tx = await wallet.executeMetaTx(methodData, sigs, nonce, 0, gasLimit, ethers.constants.AddressZero, feeWallet.address);       
+        console.log(tx);
+        console.log(tx.logs[2].args);
         //console.log(web3utils.hexToAscii(tx.logs[0].args.returnData))
 
         newBalance = await web3.eth.getBalance(wallet2);
