+++ test/upgrade_test.js
@@ -6,49 +6,50 @@ const EnsResolver = artifacts.require("EnsResolver");
 const namehash = require('eth-ens-namehash');
 //const ENS = require('@ensdomains/ens/build/contracts/ENS.json');
 const ENS = artifacts.require("@ensdomains/ens/ENSRegistry");
-const walletLib = require("../lib/wallet.js");
 
 contract("Upgrade", accounts => {
-	var relayerWallet = web3.eth.accounts.create();
-	var feeAmount = web3.utils.toWei("0.0012345", "ether");		
-	const resolverAddr = "0x9590030b26dE3A037Cd679b33A177A645BFaC114";
-
-    async function createFactoryWallet(owner, salt) {
+	
+	it("should be able to upgrade", async () => {
+		
 		const blockNumber = await web3.eth.getBlockNumber();
+		const resolverAddr = "0x9590030b26dE3A037Cd679b33A177A645BFaC114";
 
 		var merkelHeight = 6;
-		var dailyLimit = web3.utils.toWei("1000");
+		var salt = 100;
+		var dailyLimit = web3.utils.toWei("0.01", "ether");
 		var walletFactory = await commons.createWalletFactory(resolverAddr);
-		var {root_arr, leaves_arr} = walletLib.createHOTP("SECRET", merkelHeight);
+		var tmpWallet = web3.eth.accounts.create();
+		var {root_arr, leaves_arr} = commons.createHOTP(merkelHeight);
+		var relayerWallet = web3.eth.accounts.create();
 		var feeReceipient = relayerWallet.address;
+		var feeAmount = web3.utils.toWei("0.0012345", "ether");		
+		var sigs = await commons.signCreateTx([tmpWallet], tmpWallet.address, root_arr, merkelHeight, tmpWallet.address, dailyLimit, salt, feeReceipient, feeAmount);
 
-        const walletAddrComputed = await walletFactory.computeWalletAddress(
-			owner,
+		const walletAddrComputed = await walletFactory.computeWalletAddress(
+			tmpWallet.address,
 			salt
 		  );
 		await web3.eth.sendTransaction({ from: accounts[0], to: walletAddrComputed, value: web3.utils.toWei("2", "ether") , gas: 300000});
+		var newBalance = await web3.eth.getBalance(walletAddrComputed);
+		console.log("WalltComputed", newBalance);		
 
-        var subdomain = "superlongcrazynameverycheap000001" + blockNumber + salt;
+		var subdomain = "superlongcrazynameverycheap000001" + blockNumber;
 		var smartWallet = await walletFactory.createWallet({
 			resolver: resolverAddr,
 			domain: [subdomain, "crazy"],
-			owner: owner,
+			owner: tmpWallet.address,
 			rootHash: root_arr,
 			merkelHeight: merkelHeight,
-			drainAddr: owner,
+			drainAddr: tmpWallet.address,
 			dailyLimit: dailyLimit,
 			salt: salt,
 			feeReceipient: feeReceipient,
 			feeAmount: feeAmount
 		});
+	
+		const resolver = await EnsResolver.at(resolverAddr);
+		console.log("RESOLVED!", await resolver.addr(namehash.hash(subdomain + '.crazy.one')), walletAddrComputed);
 
-        return walletAddrComputed;        
-    }
-
-	it("should be able to upgrade", async () => {
-        const blockNumber = await web3.eth.getBlockNumber();
-		var tmpWallet = web3.eth.accounts.create();
-		var walletAddrComputed = await createFactoryWallet(tmpWallet.address, blockNumber);
 		console.log(walletAddrComputed);
 		var wallet = await commons.walletWithAddress(walletAddrComputed);
 		var owner = await wallet.getOwner();
@@ -90,28 +91,29 @@ contract("Upgrade", accounts => {
 		var feeWallet = web3.eth.accounts.create();
 		methodData = wallet.contract.methods.makeTransfer(dest.address, web3.utils.toWei("0.0012345", "ether")).encodeABI();
                 
-		var sigs = await commons.signOffchain2(
-				[tmpWallet],
-				wallet.address,
-				0,
-				methodData,
-				0,
-				nonce,
-				Number(web3.utils.toWei("1", "gwei")),
-				gasLimit,
-				ethers.constants.AddressZero,
-				feeWallet.address
-		);
+        var sigs = await commons.signOffchain2(
+            [tmpWallet],
+            wallet.address,
+            0,
+            methodData,
+            0,
+            nonce,
+            Number(web3.utils.toWei("1", "gwei")),
+            gasLimit,
+            ethers.constants.AddressZero,
+            feeWallet.address
+        );
 
-		const tx = await wallet.executeMetaTx(methodData, sigs, nonce, Number(web3.utils.toWei("1", "gwei")), gasLimit, ethers.constants.AddressZero, feeWallet.address);
+        const tx = await wallet.executeMetaTx(methodData, sigs, nonce, Number(web3.utils.toWei("1", "gwei")), gasLimit, ethers.constants.AddressZero, feeWallet.address);
 		console.log("method=", methodData);
-		newBalance = await web3.eth.getBalance(dest.address);
-		console.log("Balance=", newBalance);
-		assert.equal(newBalance, web3.utils.toWei(".0012345", "ether"), "withdraw amount is correct");
+        newBalance = await web3.eth.getBalance(dest.address);
+        console.log("Balance=", newBalance);
+        assert.equal(newBalance, web3.utils.toWei(".0012345", "ether"), "withdraw amount is correct");
 
 		newBalance = await web3.eth.getBalance(feeWallet.address);
 		console.log("Fee Balance=", newBalance, "Actual cost=", tx.receipt.gasUsed);
 		assert.isTrue(newBalance > tx.receipt.gasUsed);
 
+		console.log(await walletFactory.getCreated());
 	});
 });
