+++ truffle-config.js
@@ -56,7 +56,7 @@ module.exports = {
         host: "127.0.0.1",
         port: 8545, // port: 7545 for Ganashe
         network_id: "*", // Match any network id
-        gas: 6000000,
+        gas: 4000000,
       },
       advanced: {
         port: 8777,             // Custom port
@@ -127,13 +127,13 @@ module.exports = {
   
     // Set default mocha options here, use special reporters etc.
     mocha: {
-      timeout: 30000
+      // timeout: 100000
     },
   
     // Configure your compilers
     compilers: {
       solc: {
-        version: "0.8.0",    // Fetch exact version from solc-bin (default: truffle's version)
+        version: "0.7.6",    // Fetch exact version from solc-bin (default: truffle's version)
         // docker: true,        // Use "0.5.1" you've installed locally with docker (default: false)
         settings: {          // See the solidity docs for advice about optimization and evmVersion
          optimizer: {
